+++ examples/sim/src/server.hpp
@@ -45,7 +45,8 @@ private:
   void locoEStop(uint16_t loco_addr) final;
   void locoPurge(uint16_t loco_addr) final;
   [[nodiscard]] z21::LocoInfo locoInfo(uint16_t loco_addr) final;
-  void locoName(uint16_t loco_addr, uint8_t index, std::string_view name) final;
+  [[nodiscard]] z21::LocoEntry locoEntry(uint16_t loco_addr) final;
+  void locoEntry(uint16_t loco_addr, z21::LocoEntry loco_entry) final;
   void locoDrive(uint16_t loco_addr,
                  z21::LocoInfo::SpeedSteps speed_steps,
                  uint8_t rvvvvvvv) final;